Use Babel Plugin Import Glob Array
A small, simple babel plugin that will convert any import path
containing a glob pattern as an array of each of the files that match
the pattern.
We use it to import all *.mdx files with their metadata to contruct the
section navigation.

Source code https://github.com/jescalan/babel-plugin-import-glob-array