Associate homepage page title and heading
Adjust title and main heading to create association for screenreader
users.

“The page title was reported by JAWS as ‘GOV.UK Platform as a Service’.
However, the level 1 heading on the homepage is ‘Host your service in
the cloud’. This is confusing because it is unclear as to what the page
refers. When combined, page titles and level 1 headings describe the
purpose and function of a page and its content. I would therefore expect
he page title and level 1 heading to match.”

Ticket: https://www.pivotaltracker.com/n/projects/1275640/stories/174314818