Generate page contents list
Problem:

DAC audit found that

An underlined hash is displayed with mouse hover of the level-two
headings on the ‘GOV.UK Platform as a Service’ page. The purpose of the
hash is unclear as this only appears to direct users’ focus to the
adjacent heading when selected. Also, they are not accessible to
standard keyboard commands.

We implemented that feature to allow us to provide deep-links in
engagement.

Solution:

Instead on providing this functionality on mouse hover only, we now
generate a page table of contents based on headings' IDs, that are
anchor links to them. This way we retain the functionality and also
make it accessible to everyone.

To generate table of content's we're using rehype-toc plugin.
This generates a TOC on every page, which is undesirable as we don't
need one on the homepage, the 404 page, the /sign-in page or any page
that doesn't have headings (like /roadmap).

To remove the TOC from those page, we've created a rehype-plugin. We
could have created a more comprehensive, own TOC plugin with page
exclusions, but that would have taken longer.