Fix failing build for optional props
To customise the footer items and sections, we use a JSON config file.

We import that file into the component view and populate.
Not all properties are required, but when omitted, Typescript will throw
an error.

To mitigate, we now list all properties in the JSON file, but specify a
`null` or empty values.

Additionally checks if the view have been updated so that they don't
output any empty markup.