Add a cookie-banner component
Create a react component and import it into the template.

I opted to to go with a link to the cookies page to manage the consent
instead of the yes/no button.

We're going to implement a similar banner on other .cloud.service
subdomains but consent management will always be in here and wouldn't
want to set expectations.

It also gives users a change to see what cookies they're looking to
accept.