Less verbose Server header
This reduces the information server responds with to just the vendor
instead of vendor + version number as it does now.

Removing all header information relating to the nginx installation
requires the installation and use of third-party modules, namely
“nginx-module-security-headers” or, “nginx-module-headers-more”.

As we're using the nginx buildpack where these aren't complied,
we're going to raise a feature request.