Add missing security headers
The X Frame Options header prevents a site from loading
iframes, preventing possible “Clickjacking” attacks. Clickjacking
is an attack where seemingly-benign content is reproduced
within the context of a site controlled by a malicious party.
Through using transparent or invisible content, a user appears
to be interacting with trusted content, yet actually be
undertaking completely different actions.

The X XSS Protection header is designed to detect and assist in
protecting against cross-site scripting (“XSS”) attacks. However,
most modern browsers ignore this header now as having it on
can introduce additional security issues. The header should
therefore be set to off.

The X Content Type header is intended to prevent a browser
from sniffing a response away from the declared content type
of a specified resource. It assists in reducing the possibility of a
drive-by download attack. In such an attack, a user may
assume that they are downloading a legitimate and trusted
resource. An appropriately-placed attacker may be able to
intercept such a request, and replace the original resource
with malicious content.