Implement a Content Security Policy
The Content Security Policy (“CSP”) header defines content
sources that are trusted and approved by the browser to load
specific resources such as scripts. The header also defines how
loaded resources may behave. The header assists in defending
against cross-site scripting (“XSS”) and other code execution
attacks.

An effective CSP requires a way to validate sources.
When it comes to script source, there are a couple of ways to do it -
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src
- and we've chosen to use a `nonce` value that is unique to each request.

As these are static pages served by nginx buildpack -
https://docs.cloudfoundry.org/buildpacks/nginx/index.html -
we're limited in ways we can generate and inject a nonce value.

We utilise the `ngx_http_sub_module` that is compiled in the nginx
buildpack to update the **CSP_NONCE_VAL** in the HTML markup with a
generated value at request time

To generate that value we use nginx's $request_Id property, which is
randomly generated string of 32 hexadecimal characters.

We finanly add a Content-Security-Policy header where we specify which
sources are allowed and that script-src need a nonce- value.
Specifying nonce makes a modern browser ignore 'unsafe-inline' which is
set for older browsers without nonce support.