Add event tracking analytics functions
To measure engagement for users that have accepted cookies we want to
record certain events with either specific labels for targeted elements
or more generic for larger groups .

This script introduces functions to send events for:

- targeted elements with data attributes
- external links to non *.cloud.service subdomains and email links
- 404 pages

For targeted elements values for teh event are read from the data
attributes.

For external links, values are partially set and partially read from
the target element.

For 404 pages, values are set.