Update Rollup and app to serve hashed CSS&JS
While Nextjs can easily bundle and hash a CSS file, it cannot do the
same for any client-side javascript bundle that need is executed via
`<script>`

To work around this, we've updated the Rollup config to hash both
CSS and JS files and used a plugin to generate a `manifest.json`
file for each.

In App and Document we then import those files and use hashed names
to reference the correct paths.

To make sure we alwats have a JSON file available for import, we
updated the `local` npm task so that rollup buidl is run first and
then next dev with rollup:watch.