Remove tenant logos
This removes tenant logos displayed on the homepage as well as source
files and CSS snippets.

We have lots of tenants but only display a few JPEG logos, which next
to the text bring very little value.

We could later look at showin the list in the admin interface on
the /performance page.

Additionally this change will save us 5 HTTP request and ~46KB in page
weight.